+++ src/gmp.c
@@ -13259,11 +13259,14 @@ handle_get_info (gmp_parser_t *gmp_parser, GError **error)
                              "<score>%d</score>"
                              "<cve_refs>%s</cve_refs>"
                              "<status>%s</status>",
-                             cpe_info_iterator_nvd_id (&info),
+                             cpe_info_iterator_nvd_id (&info)
+                              ? cpe_info_iterator_nvd_id (&info)
+                              : "",
                              cpe_info_iterator_score (&info),
                              cpe_info_iterator_cve_refs (&info),
-                             cpe_info_iterator_status (&info) ?
-                             cpe_info_iterator_status (&info) : "");
+                             cpe_info_iterator_status (&info)
+                              ? cpe_info_iterator_status (&info)
+                              : "");
 
           if (get_info_data->details == 1)
             {
@@ -17914,6 +17917,7 @@ handle_get_vulns (gmp_parser_t *gmp_parser, GError **error)
                                "<creation_time>%s</creation_time>"
                                "<modification_time>%s</modification_time>"
                                "<severity>%1.1f</severity>"
+                               "<score>%i</score>"
                                "<qod>%d</qod>",
                                get_iterator_uuid (&vulns),
                                get_iterator_name (&vulns),
@@ -17921,6 +17925,7 @@ handle_get_vulns (gmp_parser_t *gmp_parser, GError **error)
                                get_iterator_creation_time (&vulns),
                                get_iterator_modification_time (&vulns),
                                vuln_iterator_severity (&vulns),
+                               vuln_iterator_score (&vulns),
                                vuln_iterator_qod (&vulns));
 
       // results for the vulnerability