+++ src/manage_migrators.c
@@ -2533,6 +2533,39 @@ migrate_238_to_239 ()
   return 0;
 }
 
+/**
+ * @brief Migrate the database from version 239 to version 240.
+ *
+ * @return 0 success, -1 error.
+ */
+int
+migrate_239_to_240 ()
+{
+  sql_begin_immediate ();
+
+  /* Ensure that the database is currently version 239. */
+
+  if (manage_db_version () != 239)
+    {
+      sql_rollback ();
+      return -1;
+    }
+
+  /* Update the database. */
+
+  /* Table results also got a score column, for extended severities. */
+
+  sql ("DROP VIEW vulns;");
+
+  /* Set the database version to 240. */
+
+  set_db_version (240);
+
+  sql_commit ();
+
+  return 0;
+}
+
 #undef UPDATE_DASHBOARD_SETTINGS
 
 /**
@@ -2578,6 +2611,7 @@ static migrator_t database_migrators[] = {
   {237, migrate_236_to_237},
   {238, migrate_237_to_238},
   {239, migrate_238_to_239},
+  {240, migrate_239_to_240},
   /* End marker. */
   {-1, NULL}};